Upgraded Service Management Versioning to 2014-04-01

+++ lib/azure/virtual_machine_management/virtual_machine_management_service.rb
@@ -51,7 +51,7 @@ module Azure
       #
       # Returns an  Azure::VirtualMachineManagement::VirtualMachine instance.
       def get_virtual_machine(name, cloud_service_name)
-        server = list_virtual_machines(cloud_service_name).select { |x| x.vm_name == name && x.cloud_service_name == cloud_service_name }
+        server = list_virtual_machines(cloud_service_name).select { |x| x.vm_name == name.downcase }
         server.first
       end
 
@@ -103,63 +103,104 @@ module Azure
       # See:
       # http://msdn.microsoft.com/en-us/library/windowsazure/jj157194.aspx
       # http://msdn.microsoft.com/en-us/library/windowsazure/jj157186.aspx
-      def create_virtual_machine(params, options = {}, add_role = false)
+      def create_virtual_machine(params, options = {})
         options[:os_type] = get_os_type(params[:image])
         validate_deployment_params(params, options)
         options[:deployment_name] ||= options[:cloud_service_name]
-
-        unless add_role
-          Loggerx.info 'Creating deploymnent...'
-          options[:cloud_service_name] ||= generate_cloud_service_name(params[:vm_name])
-          options[:storage_account_name] ||= generate_storage_account_name(params[:vm_name])
-          optionals = {}
-          if options[:virtual_network_name]
-            virtual_network_service = Azure::VirtualNetworkManagementService.new
-            virtual_networks = virtual_network_service.list_virtual_networks.select { |x| x.name == options[:virtual_network_name] }
-            if virtual_networks.empty?
-              Loggerx.error_with_exit "Virtual network #{options[:virtual_network_name]} doesn't exists"
-            else
-              optionals[:affinity_group_name] = virtual_networks.first.affinity_group
-            end
-          elsif options[:affinity_group_name]
-            optionals[:affinity_group_name] = options[:affinity_group_name]
+        Loggerx.info 'Creating deploymnent...'
+        options[:cloud_service_name] ||= generate_cloud_service_name(params[:vm_name])
+        options[:storage_account_name] ||= generate_storage_account_name(params[:vm_name])
+        optionals = {}
+        if options[:virtual_network_name]
+          virtual_network_service = Azure::VirtualNetworkManagementService.new
+          virtual_networks = virtual_network_service.list_virtual_networks.select { |x| x.name == options[:virtual_network_name] }
+          if virtual_networks.empty?
+            Loggerx.error_with_exit "Virtual network #{options[:virtual_network_name]} doesn't exists"
           else
-            optionals[:location] = params[:location]
+            optionals[:affinity_group_name] = virtual_networks.first.affinity_group
           end
-          cloud_service = Azure::CloudServiceManagementService.new
-          cloud_service.create_cloud_service(options[:cloud_service_name], optionals)
-          cloud_service.upload_certificate(options[:cloud_service_name], params[:certificate]) unless params[:certificate].empty?
-          Azure::StorageManagementService.new.create_storage_account(options[:storage_account_name], optionals)
-          body = Serialization.deployment_to_xml(params, options)
-          path = "/services/hostedservices/#{options[:cloud_service_name]}/deployments"
+        elsif options[:affinity_group_name]
+          optionals[:affinity_group_name] = options[:affinity_group_name]
         else
-          cloud_services = Azure::CloudServiceManagementService.new.get_cloud_service_properties(
-            options[:cloud_service_name]
-          )
-          existing_ports = []
-          # There should be only one cloud_serivce in the Array.
-          cloud_services.each do |cloud_service|
-            cloud_service.virtual_machines[options[:deployment_name].to_sym].each do |vm|
-              vm.tcp_endpoints.each do |endpoint|
-                existing_ports << endpoint[:public_port]
-              end
-            end
-          end
-
-          options[:existing_ports] = existing_ports
-
-          Loggerx.info 'Deployment exists, adding role...'
-          body = Serialization.role_to_xml(params, options).to_xml
-          path = "/services/hostedservices/#{options[:cloud_service_name]}/deployments/#{options[:deployment_name]}/roles"
+          optionals[:location] = params[:location]
         end
-        Loggerx.error 'Cloud service name is required for adding role.' unless options[:cloud_service_name]
-        Loggerx.error 'Storage account name is required for adding role.' unless options[:storage_account_name]
+        cloud_service = Azure::CloudServiceManagementService.new
+        cloud_service.create_cloud_service(options[:cloud_service_name], optionals)
+        cloud_service.upload_certificate(options[:cloud_service_name], params[:certificate]) unless params[:certificate].empty?
+        Azure::StorageManagementService.new.create_storage_account(options[:storage_account_name], optionals)
+        body = Serialization.deployment_to_xml(params, options)
+        path = "/services/hostedservices/#{options[:cloud_service_name]}/deployments"
         Loggerx.info 'Deployment in progress...'
         request = ManagementHttpRequest.new(:post, path, body)
         request.call
         get_virtual_machine(params[:vm_name], options[:cloud_service_name])
       rescue Exception => e
         e.message
+      end
+
+      # Public: Add a new role to a cloud service. Atleast one deployment should exist before you can add a role.
+      #
+      # ==== Attributes
+      #
+      # * +params+    - Hash.  parameters.
+      # * +options+   - Hash.  Optional parameters.
+      #
+      #  ==== Params
+      #
+      # Accepted key/value pairs are:
+      # * +:vm_name+             - String.  Name of virtual machine.
+      # * +:vm_user+             - String.  User name for the virtual machine instance.
+      # * +:password+            - String.  A description for the hosted service.
+      # * +:image+               - String.  Name of the disk image to use to create the virtual machine.
+      # * +:cloud_service_name+  - String. Name of cloud service.
+      #
+      #  ==== Options
+      #
+      # Accepted key/value pairs are:
+      # * +:storage_account_name+  - String. Name of storage account.
+      # * +:tcp_endpoints+         - String. Specifies the internal port and external/public port separated by a colon.
+      #   You can map multiple internal and external ports by separating them with a comma.
+      # * +:ssh_private_key_file+  - String. Path of private key file.
+      # * +:ssh_certificate_file+  - String. Path of certificate file.
+      # * +:ssh_port+              - Integer. Specifies the SSH port number.
+      # * +:winrm_http_port        - Integer. Specifies the WinRM HTTP port number.
+      # * +:winrm_https_port       - Integer. Specifies the WinRM HTTPS port number.
+      # * +:vm_size+               - String. Specifies the size of the virtual machine instance.
+      # * +:winrm_transport+       - Array. Specifies WINRM transport protocol.
+      #
+      # Returns Azure::VirtualMachineManagement::VirtualMachine objects of newly created instance.
+      #
+      # See:
+      # http://msdn.microsoft.com/en-us/library/windowsazure/jj157186.aspx
+      def add_role(params, options = {})
+        options[:os_type] = get_os_type(params[:image])
+        validate_deployment_params(params, options, true)
+        cloud_service = Azure::CloudServiceManagementService.new
+        cloud_service = cloud_service.get_cloud_service_properties(params[:cloud_service_name])
+        deployment_name = cloud_service.deployment_name
+        Loggerx.error_with_exit "Deployment doesn't exists." if cloud_service && deployment_name.empty?
+        others = {}
+        if cloud_service.location
+          others[:location] = cloud_service.location
+        elsif cloud_service.affinity_group
+          others[:affinity_group_name] = cloud_service.affinity_group
+        end
+        options[:storage_account_name] ||= generate_storage_account_name(params[:vm_name])
+        Azure::StorageManagementService.new.create_storage_account(options[:storage_account_name], others)
+        Loggerx.info 'Deployment exists, adding role...'
+        existing_ports = []
+        cloud_service.virtual_machines[deployment_name.to_sym].each do |vm|
+          vm.tcp_endpoints.each do |endpoint|
+            existing_ports << endpoint[:public_port]
+          end
+        end
+        options[:existing_ports] = existing_ports
+        body = Serialization.role_to_xml(params, options).to_xml
+        path = "/services/hostedservices/#{cloud_service.name}/deployments/#{deployment_name}/roles"
+        Loggerx.info 'Deployment in progress...'
+        request = ManagementHttpRequest.new(:post, path, body)
+        request.call
+        get_virtual_machine(params[:vm_name], cloud_service.name)
       end
 
       # Public: Deletes the deployment, cloud service and disk.
@@ -174,11 +215,19 @@ module Azure
       #
       # Returns NONE
       def delete_virtual_machine(vm_name, cloud_service_name)
-        vm = get_virtual_machine(vm_name, cloud_service_name)
+        virtual_machines = list_virtual_machines(cloud_service_name)
+        vm = virtual_machines.select { |x| x.vm_name == vm_name }.first
         if vm
-          cloud_service = Azure::CloudServiceManagementService.new
-          cloud_service.delete_cloud_service_deployment(cloud_service_name)
-          cloud_service.delete_cloud_service(cloud_service_name)
+          if virtual_machines.size == 1
+            cloud_service = Azure::CloudServiceManagementService.new
+            cloud_service.delete_cloud_service_deployment(cloud_service_name)
+            cloud_service.delete_cloud_service(cloud_service_name)
+          else
+            path = "/services/hostedservices/#{vm.cloud_service_name}/deployments/#{vm.deployment_name}/roles/#{vm.vm_name}"
+            Loggerx.info "Deleting virtual machine #{vm_name}. \n"
+            request = ManagementHttpRequest.new(:delete, path)
+            request.call
+          end
           Loggerx.info "Waiting for disk to be released.\n"
           disk_name = vm.disk_name
           disk_management_service = VirtualMachineDiskManagementService.new
@@ -217,7 +266,7 @@ module Azure
       def shutdown_virtual_machine(vm_name, cloud_service_name)
         vm = get_virtual_machine(vm_name, cloud_service_name)
         if vm
-          if ['StoppedVM', 'StoppedDeallocated'].include?(vm.status)
+          if %w(StoppedVM StoppedDeallocated).include?(vm.status)
             Loggerx.error 'Cannot perform the shutdown operation on a stopped virtual machine.'
           elsif vm.deployment_status == 'Running'
             path = "/services/hostedservices/#{vm.cloud_service_name}/deployments/#{vm.deployment_name}/roleinstances/#{vm.vm_name}/Operations"
@@ -282,7 +331,7 @@ module Azure
           Loggerx.error "Cannot find virtual machine \"#{vm_name}\" under cloud service \"#{cloud_service_name}\"."
         end
       end
-      
+
       # Public: Add/Update endpoints of virtual machine.
       #
       # ==== Attributes
@@ -304,7 +353,7 @@ module Azure
       # * +:protocol+              - String. Specifies the transport protocol
       #   for the endpoint. Possible values are: TCP, UDP
       # * +:direct_server_return+  - String. Specifies whether the endpoint
-      #   uses Direct Server Return. (optional)
+      #   uses Direct Server Return. Possible values are: true, false (optional)
       # * +:load_balancer          - Hash. Contains properties that define the
       #   endpoint settings that the load balancer uses to monitor the
       #   availability of the Virtual Machine (optional)
@@ -333,7 +382,7 @@ module Azure
           path = "/services/hostedservices/#{vm.cloud_service_name}/deployments/#{vm.deployment_name}/roles/#{vm_name}"
           endpoints = vm.tcp_endpoints + vm.udp_endpoints
           input_endpoints.each do |iep|
-            endpoints.delete_if { |ep| iep[:name].downcase == ep[:name].downcase && iep[:protocol].downcase  == ep[:protocol] }
+            endpoints.delete_if { |ep| iep[:name].downcase == ep[:name].downcase }
           end
           endpoints += input_endpoints
           body = Serialization.update_role_to_xml(endpoints, vm)
@@ -352,7 +401,7 @@ module Azure
       # * +name+                - String. Virtual machine name.
       # * +cloud_service_name+  - String. Cloud service name.
       # * +endpoint_name+       - String. Name of endpoint.
-      # 
+      #
       # See http://msdn.microsoft.com/en-us/library/windowsazure/jj157187.aspx
       #
       # Returns NONE
@@ -377,8 +426,6 @@ module Azure
       #
       # * +cloud_service_name+  - String. Cloud service name.
       # * +vm_name+           - String. Virtual machine name.
-      # * +lun+                 - String. Specifies the Logical Unit Number
-      #   (LUN) for the disk. Valid LUN values are 0 through 15.
       # * +options+             - Hash. Optional parameters.
       #
       # ==== Options
@@ -396,12 +443,12 @@ module Azure
       # See http://msdn.microsoft.com/en-us/library/windowsazure/jj157199.aspx
       #
       # Returns None
-      def add_data_disk(vm_name, cloud_service_name, lun, options = {})
+      def add_data_disk(vm_name, cloud_service_name, options = {})
         options[:import] ||= false
         vm = get_virtual_machine(vm_name, cloud_service_name)
         if vm
           path = "/services/hostedservices/#{cloud_service_name}/deployments/#{vm.deployment_name}/roles/#{vm_name}/DataDisks"
-          body = Serialization.add_data_disk_to_xml(lun, vm.media_link, options)
+          body = Serialization.add_data_disk_to_xml(vm, options)
           Loggerx.info "Adding data disk to virtual machine #{vm_name} ..."
           request = ManagementHttpRequest.new(:post, path, body)
           request.call
@@ -430,15 +477,17 @@ module Azure
         random_string(vm_name + 'storage').gsub(/[^0-9a-z ]/i, '').downcase[0..23]
       end
 
-      def validate_deployment_params(params, options)
+      def validate_deployment_params(params, options, add_role = false)
         errors = []
-        params_keys = ['vm_name', 'image', 'location', 'vm_user']
-        if options[:os_type] == 'Windows'
-          params_keys += ['password']
-        end
+        params_keys = %w(vm_name image vm_user)
+        params_keys += ['password'] if options[:os_type] == 'Windows'
         options_keys = []
-        options_keys = ['private_key_file', 'certificate_file'] if certificate_required?(params, options)
-
+        options_keys = %w(private_key_file certificate_file) if certificate_required?(params, options)
+        if add_role
+          params_keys += ['cloud_service_name']
+        else
+          params_keys += ['location']
+        end
         params_keys.each do |key|
           errors << key if params[key.to_sym].nil?
         end
@@ -446,9 +495,10 @@ module Azure
         options_keys.each do |key|
           errors << key if options[key.to_sym].nil?
         end
-        validate_role_size(options[:vm_size])
-        validate_location(params[:location]) unless errors.include?('location')
+
         if errors.empty?
+          validate_location(params[:location]) unless add_role
+          validate_role_size(options[:vm_size])
           params[:certificate] = {}
           if certificate_required?(params, options)
             begin
@@ -479,7 +529,7 @@ module Azure
       end
 
       def validate_role_size(vm_size)
-        valid_role_sizes = ['ExtraSmall', 'Small', 'Medium', 'Large', 'ExtraLarge', 'A6', 'A7']
+        valid_role_sizes = %w(ExtraSmall Small Medium Large ExtraLarge A5 A6 A7 Basic_A0 Basic_A1 Basic_A2 Basic_A3 Basic_A4)
         if vm_size && !valid_role_sizes.include?(vm_size)
           Loggerx.error_with_exit "Value '#{vm_size}' specified for parameter 'vm_size' is invalid. Allowed values are 'ExtraSmall,Small,Medium,Large,ExtraLarge,A6,A7'"
         end
